refactor: introduce sub commands
